This deals with most of #3.
Instead of a baseline being a simple integer as before, there are now 3
different kinds of referencing baselines in GRID_LAYOUT positions: an
absolute reference, which corresponds to what was available before, an
automatic baseline, which is computed based on the height of the
component, and a relative reference, which is a function of the start
and/or end baselines of a given set of components. During rendering the
baseline dependencies are resolved through a topological sort.

Currently GRID_LAYOUT still only deals with a viewport of fixed width
and height, and what is left to do is to relax this requirement to
allow flexible height.